requirements.txt
RPi.GPIO
paho.mqtt
python-opencv
pillow
opencv-contrib-python
flask
IPy
pypinyin